+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
+- Convert the first unit test to nimut/testing-framework (#194)
 - Move to old tests to the "Legacy" namespace (#193)
 
 ### Deprecated